# Conflicts:
#	devoir2/Unishop/src/App.java
#	devoir2/Unishop/src/Data/Entities/Products/LearningResource.java
#	devoir2/Unishop/src/Data/Entities/Users/Client.java
#	devoir2/Unishop/src/Data/Entities/Users/User.java

+++ devoir2/Unishop/src/Data/Entities/Products/Product.java
@@ -68,7 +68,7 @@ public class Product {
         return quantity;
     }
 
-    public void setquantity(int quantity) {
+    public void setquantity(long quantity) {
         this.quantity = quantity;
     }
 
@@ -84,7 +84,7 @@ public class Product {
         return points;
     }
 
-    public void setPoints(int points) {
+    public void setPoints(long points) {
         this.points = points;
     }
